+++ cocos2d/core/assets/index.js
@@ -41,5 +41,4 @@ require('./CCLabelAtlas');
 require('./CCTextAsset');
 require('./CCJsonAsset');
 require('./CCBufferAsset');
-require('./CCEffectAsset');
 require('./material');
