+++ cocos2d/core/assets/material/index.js
@@ -1,2 +1,3 @@
-import './material-variant';
+import './CCEffectAsset';
 import './CCMaterial';
+import './material-variant';
