+++ cocos2d/tilemap/CCTiledLayer.js
@@ -28,6 +28,7 @@ const Material = require('../core/assets/material/CCMaterial');
 const RenderFlow = require('../core/renderer/render-flow');
 
 import { Mat4, Vec2 } from '../core/value-types';
+import MaterialVariant from '../core/assets/material/material-variant';
 let _mat4_temp = cc.mat4();
 let _vec2_temp = cc.v2();
 let _vec2_temp2 = cc.v2();
@@ -1289,20 +1290,19 @@ let TiledLayer = cc.Class({
 
             let material = this.sharedMaterials[i];
             if (!material) {
-                material = Material.getInstantiatedBuiltinMaterial('2d-sprite', this);
+                material = Material.getBuiltinMaterial('2d-sprite');
             }
-            else {
-                material = Material.getInstantiatedMaterial(material, this);
-            }
+            material = MaterialVariant.create(material, this);
 
             material.define('CC_USE_MODEL', true);
             material.setProperty('texture', texture);
-            this.setMaterial(i, material);
+
+            this.sharedMaterials[i] = material;
+            
             texIdMatIdx[tilesetIdx] = i;
         }
 
-        this.markForRender(true);
-    },
+    }
 });
 
 cc.TiledLayer = module.exports = TiledLayer;
